infrastructure/ansible/role_service-host/scripts/requirements.txt

## Storage
# S3
minio == 7.1.14 

## General
python-dotenv == 1.0.0 
